chore(PSDK-540): Move test factories under a factories package
