+++ tests/test_address.py
@@ -6,46 +6,31 @@ import pytest
 from cdp.address import Address
 from cdp.balance_map import BalanceMap
 from cdp.client.exceptions import ApiException
-from cdp.client.models.asset import Asset as AssetModel
-from cdp.client.models.balance import Balance as BalanceModel
 from cdp.errors import ApiError
 from cdp.faucet_transaction import FaucetTransaction
 
 
-@pytest.fixture
-def address():
-    """Create and return a fixture for an Address."""
-    return Address(
-        network_id="base-sepolia", address_id="0x1234567890123456789012345678901234567890"
-    )
-
-
-@pytest.fixture
-def asset_model():
-    """Create and return a fixture for an AssetModel."""
-    return AssetModel(network_id="base-sepolia", asset_id="eth", decimals=18)
-
-
-@pytest.fixture
-def balance_model(asset_model):
-    """Create and return a fixture for a BalanceModel."""
-    return BalanceModel(amount="1000000000000000000", asset=asset_model)
-
-
-def test_address_initialization(address):
+def test_address_initialization(address_factory):
     """Test the initialization of an Address."""
+    address = address_factory()
+
+    assert isinstance(address, Address)
     assert address.network_id == "base-sepolia"
     assert address.address_id == "0x1234567890123456789012345678901234567890"
 
 
-def test_address_can_sign(address):
+def test_address_can_sign(address_factory):
     """Test the can_sign property of an Address."""
+    address = address_factory()
+
     assert not address.can_sign
 
 
 @patch("cdp.Cdp.api_clients")
-def test_address_faucet(mock_api_clients, address):
+def test_address_faucet(mock_api_clients, address_factory):
     """Test the faucet method of an Address."""
+    address = address_factory()
+
     mock_request_faucet = Mock()
     mock_request_faucet.return_value = Mock(spec=FaucetTransaction)
     mock_api_clients.external_addresses.request_external_faucet_funds = mock_request_faucet
@@ -59,8 +44,10 @@ def test_address_faucet(mock_api_clients, address):
 
 
 @patch("cdp.Cdp.api_clients")
-def test_address_faucet_with_asset_id(mock_api_clients, address):
+def test_address_faucet_with_asset_id(mock_api_clients, address_factory):
     """Test the faucet method of an Address with an asset_id."""
+    address = address_factory()
+
     mock_request_faucet = Mock()
     mock_request_faucet.return_value = Mock(spec=FaucetTransaction)
     mock_api_clients.external_addresses.request_external_faucet_funds = mock_request_faucet
@@ -74,8 +61,10 @@ def test_address_faucet_with_asset_id(mock_api_clients, address):
 
 
 @patch("cdp.Cdp.api_clients")
-def test_address_faucet_api_error(mock_api_clients, address):
+def test_address_faucet_api_error(mock_api_clients, address_factory):
     """Test the faucet method of an Address raises an error when the API call fails."""
+    address = address_factory()
+
     mock_request_faucet = Mock()
     err = ApiException(500, "boom")
     mock_request_faucet.side_effect = ApiError(err, code="boom", message="boom")
@@ -86,8 +75,11 @@ def test_address_faucet_api_error(mock_api_clients, address):
 
 
 @patch("cdp.Cdp.api_clients")
-def test_address_balance(mock_api_clients, address, balance_model):
+def test_address_balance(mock_api_clients, address_factory, balance_model_factory):
     """Test the balance method of an Address."""
+    address = address_factory()
+    balance_model = balance_model_factory()
+
     mock_get_balance = Mock()
     mock_get_balance.return_value = balance_model
     mock_api_clients.external_addresses.get_external_address_balance = mock_get_balance
@@ -102,8 +94,10 @@ def test_address_balance(mock_api_clients, address, balance_model):
 
 
 @patch("cdp.Cdp.api_clients")
-def test_address_balance_zero(mock_api_clients, address):
+def test_address_balance_zero(mock_api_clients, address_factory):
     """Test the balance method of an Address returns 0 when the balance is not found."""
+    address = address_factory()
+
     mock_get_balance = Mock()
     mock_get_balance.return_value = None
     mock_api_clients.external_addresses.get_external_address_balance = mock_get_balance
@@ -115,8 +109,10 @@ def test_address_balance_zero(mock_api_clients, address):
 
 
 @patch("cdp.Cdp.api_clients")
-def test_address_balance_api_error(mock_api_clients, address):
+def test_address_balance_api_error(mock_api_clients, address_factory):
     """Test the balance method of an Address raises an error when the API call fails."""
+    address = address_factory()
+
     mock_get_balance = Mock()
     err = ApiException(500, "boom")
     mock_get_balance.side_effect = ApiError(err, code="boom", message="boom")
@@ -127,8 +123,11 @@ def test_address_balance_api_error(mock_api_clients, address):
 
 
 @patch("cdp.Cdp.api_clients")
-def test_address_balances(mock_api_clients, address, balance_model):
+def test_address_balances(mock_api_clients, address_factory, balance_model_factory):
     """Test the balances method of an Address."""
+    address = address_factory()
+    balance_model = balance_model_factory()
+
     mock_list_balances = Mock()
     mock_list_balances.return_value = Mock(data=[balance_model])
     mock_api_clients.external_addresses.list_external_address_balances = mock_list_balances
@@ -145,8 +144,10 @@ def test_address_balances(mock_api_clients, address, balance_model):
 
 
 @patch("cdp.Cdp.api_clients")
-def test_address_balances_api_error(mock_api_clients, address):
+def test_address_balances_api_error(mock_api_clients, address_factory):
     """Test the balances method of an Address raises an error when the API call fails."""
+    address = address_factory()
+
     mock_list_balances = Mock()
     err = ApiException(500, "boom")
     mock_list_balances.side_effect = ApiError(err, code="boom", message="boom")
@@ -156,13 +157,17 @@ def test_address_balances_api_error(mock_api_clients, address):
         address.balances()
 
 
-def test_address_str_representation(address):
+def test_address_str_representation(address_factory):
     """Test the str representation of an Address."""
+    address = address_factory()
+
     expected_str = f"Address: (address_id: {address.address_id}, network_id: {address.network_id})"
     assert str(address) == expected_str
 
 
-def test_address_repr(address):
+def test_address_repr(address_factory):
     """Test the repr representation of an Address."""
+    address = address_factory()
+
     expected_repr = f"Address: (address_id: {address.address_id}, network_id: {address.network_id})"
     assert repr(address) == expected_repr
