+++ tests/test_transfer.py
@@ -5,123 +5,10 @@ import pytest
 from eth_account.signers.local import LocalAccount
 
 from cdp.asset import Asset
-from cdp.client.models.asset import Asset as AssetModel
-from cdp.client.models.sponsored_send import SponsoredSend as SponsoredSendModel
-from cdp.client.models.transaction import Transaction as TransactionModel
-from cdp.client.models.transfer import Transfer as TransferModel
 from cdp.errors import TransactionNotSignedError
 from cdp.sponsored_send import SponsoredSend
 from cdp.transaction import Transaction
 from cdp.transfer import Transfer
-
-
-@pytest.fixture
-def asset_model_factory():
-    """Create and return a factory for creating AssetModel fixtures."""
-
-    def _create_asset_model(network_id="base-sepolia", asset_id="usdc", decimals=6):
-        return AssetModel(network_id=network_id, asset_id=asset_id, decimals=decimals)
-
-    return _create_asset_model
-
-
-@pytest.fixture
-def asset_factory(asset_model_factory):
-    """Create and return a factory for creating Asset fixtures."""
-
-    def _create_asset(network_id="base-sepolia", asset_id="usdc", decimals=6):
-        asset_model = asset_model_factory(network_id, asset_id, decimals)
-        return Asset.from_model(asset_model)
-
-    return _create_asset
-
-
-@pytest.fixture
-def sponsored_send_model_factory():
-    """Create and return a factory for creating SponsoredSendModel fixtures."""
-
-    def _create_sponsored_send_model(status="complete"):
-        status = "submitted" if status == "broadcast" else status
-
-        return SponsoredSendModel(
-            to_address_id="0xdestination",
-            raw_typed_data="0xtypeddata",
-            typed_data_hash="0xtypeddatahash",
-            signature="0xsignature" if status in ["signed", "submitted", "complete"] else None,
-            transaction_hash="0xtransactionhash" if status == "complete" else None,
-            transaction_link="https://sepolia.basescan.org/tx/0xtransactionlink"
-            if status == "complete"
-            else None,
-            status=status,
-        )
-
-    return _create_sponsored_send_model
-
-
-@pytest.fixture
-def transaction_model_factory():
-    """Create and return a factory for creating TransactionModel fixtures."""
-
-    def _create_transaction_model(status="complete"):
-        return TransactionModel(
-            network_id="base-sepolia",
-            transaction_hash="0xtransactionhash",
-            from_address_id="0xaddressid",
-            to_address_id="0xdestination",
-            unsigned_payload="0xunsignedpayload",
-            signed_payload="0xsignedpayload"
-            if status in ["signed", "broadcast", "complete"]
-            else None,
-            status=status,
-            transaction_link="https://sepolia.basescan.org/tx/0xtransactionlink"
-            if status == "complete"
-            else None,
-        )
-
-    return _create_transaction_model
-
-
-@pytest.fixture
-def transfer_model_factory(
-    asset_model_factory, sponsored_send_model_factory, transaction_model_factory
-):
-    """Create and return a factory for creating TransferModel fixtures."""
-
-    def _create_transfer_model(gasless=True, status="complete", asset_id="usdc"):
-        asset_model = asset_model_factory(asset_id=asset_id)
-        if gasless:
-            sponsored_send_model = sponsored_send_model_factory(status)
-            transaction_model = None
-        else:
-            sponsored_send_model = None
-            transaction_model = transaction_model_factory(status)
-
-        return TransferModel(
-            network_id="base-sepolia",
-            wallet_id="test-wallet-id",
-            address_id="0xaddressid",
-            destination="0xdestination",
-            amount="1000000",  # 1 USDC or 1 ETH in wei
-            asset_id=asset_model.asset_id,
-            transfer_id="test-transfer-id",
-            asset=asset_model,
-            gasless=gasless,
-            sponsored_send=sponsored_send_model,
-            transaction=transaction_model,
-        )
-
-    return _create_transfer_model
-
-
-@pytest.fixture
-def transfer_factory(transfer_model_factory):
-    """Create and return a factory for creating Transfer fixtures."""
-
-    def _create_transfer(gasless=True, status="complete", asset_id="usdc"):
-        transfer_model = transfer_model_factory(gasless, status, asset_id)
-        return Transfer(transfer_model)
-
-    return _create_transfer
 
 
 @pytest.mark.parametrize("gasless", [True, False])
