+++ tests/test_wallet.py
@@ -5,11 +5,8 @@ import pytest
 from bip_utils import Bip32Slip10Secp256k1
 from eth_account import Account
 
-from cdp.client.models.address import Address as AddressModel
 from cdp.client.models.create_address_request import CreateAddressRequest
 from cdp.client.models.create_wallet_request import CreateWalletRequest, CreateWalletRequestWallet
-from cdp.client.models.feature_set import FeatureSet
-from cdp.client.models.wallet import Wallet as WalletModel
 from cdp.contract_invocation import ContractInvocation
 from cdp.payload_signature import PayloadSignature
 from cdp.smart_contract import SmartContract
@@ -17,93 +14,6 @@ from cdp.trade import Trade
 from cdp.transfer import Transfer
 from cdp.wallet import Wallet
 from cdp.wallet_address import WalletAddress
-
-
-@pytest.fixture
-def address_model_factory():
-    """Create and return a factory for AddressModel fixtures."""
-
-    def _create_address_model(
-        network_id="base-sepolia",
-        address_id="0x1234567890123456789012345678901234567890",
-        wallet_id="test-wallet-id",
-        public_key="0xpublickey",
-        index=0,
-    ):
-        return AddressModel(
-            network_id=network_id,
-            address_id=address_id,
-            wallet_id=wallet_id,
-            public_key=public_key,
-            index=index,
-        )
-
-    return _create_address_model
-
-
-@pytest.fixture
-def wallet_model_factory(address_model_factory):
-    """Create and return a factory for WalletModel fixtures."""
-
-    def _create_wallet_model(
-        id="test-wallet-id",
-        network_id="base-sepolia",
-        default_address=None,
-        feature_set=None,
-        server_signer_status="active_seed",
-    ):
-        if default_address is None:
-            default_address = address_model_factory()
-        if feature_set is None:
-            feature_set = FeatureSet(
-                faucet=True,
-                server_signer=True,
-                transfer=True,
-                trade=True,
-                stake=True,
-                gasless_send=True,
-            )
-        return WalletModel(
-            id=id,
-            network_id=network_id,
-            default_address=default_address,
-            feature_set=feature_set,
-            server_signer_status=server_signer_status,
-        )
-
-    return _create_wallet_model
-
-
-@pytest.fixture
-def master_key_factory():
-    """Create and return a factory for master key fixtures."""
-
-    def _create_master_key(seed=b"\x00" * 64):
-        return Bip32Slip10Secp256k1.FromSeed(seed)
-
-    return _create_master_key
-
-
-@pytest.fixture
-def wallet_factory(wallet_model_factory):
-    """Create and return a factory for Wallet fixtures."""
-
-    def _create_wallet(seed=None, **kwargs):
-        model = wallet_model_factory(**kwargs)
-        return Wallet(model, seed)
-
-    return _create_wallet
-
-
-@pytest.fixture
-def wallet_address_factory(address_model_factory):
-    """Create and return a factory for WalletAddress fixtures."""
-
-    def _create_wallet_address(key=None, **kwargs):
-        model = address_model_factory(**kwargs)
-        return WalletAddress(model, key)
-
-    return _create_wallet_address
 
 
 @patch("cdp.Cdp.use_server_signer", False)
