+++ tests/test_wallet_address.py
@@ -8,9 +8,6 @@ from eth_account.signers.local import LocalAccount
 from eth_utils import to_bytes
 from web3 import Web3
 
-from cdp.client.models.address import Address as AddressModel
-from cdp.client.models.asset import Asset as AssetModel
-from cdp.client.models.balance import Balance as BalanceModel
 from cdp.client.models.create_smart_contract_request import CreateSmartContractRequest
 from cdp.client.models.multi_token_contract_options import MultiTokenContractOptions
 from cdp.client.models.nft_contract_options import NFTContractOptions
@@ -22,90 +19,68 @@ from cdp.payload_signature import PayloadSignature
 from cdp.smart_contract import SmartContract
 from cdp.trade import Trade
 from cdp.transfer import Transfer
-from cdp.wallet_address import WalletAddress
 
 
-@pytest.fixture
-def address_model():
-    """Fixture for an AddressModel."""
-    return AddressModel(
-        network_id="base-sepolia",
-        address_id="0x1234567890123456789012345678901234567890",
-        wallet_id="test-wallet-id-1",
-        public_key="0x1234567890123456789012345678901234567890",
-        index=0,
-    )
-
-
-@pytest.fixture
-def wallet_address(address_model):
-    """Fixture for a WalletAddress."""
-    return WalletAddress(address_model)
-
-
-@pytest.fixture
-def wallet_address_with_key(address_model):
-    """Fixture for a WalletAddress with a key."""
-    key = Mock(spec=LocalAccount)
-    return WalletAddress(address_model, key)
-
-
-@pytest.fixture
-def asset_model():
-    """Fixture for an AssetModel."""
-    return AssetModel(network_id="base-sepolia", asset_id="eth", decimals=18)
-
-
-@pytest.fixture
-def balance_model(asset_model):
-    """Fixture for a BalanceModel."""
-    return BalanceModel(amount="5000000000000000000", asset=asset_model)
-
-
-def test_wallet_address_initialization(wallet_address):
+def test_wallet_address_initialization(wallet_address_factory):
     """Test the initialization of a WalletAddress."""
+    wallet_address = wallet_address_factory()
+
     assert wallet_address.network_id == "base-sepolia"
     assert wallet_address.address_id == "0x1234567890123456789012345678901234567890"
-    assert wallet_address.wallet_id == "test-wallet-id-1"
+    assert wallet_address.wallet_id == "test-wallet-id"
     assert wallet_address.key is None
     assert not wallet_address.can_sign
 
 
-def test_wallet_address_initialization_with_key(wallet_address_with_key):
+def test_wallet_address_initialization_with_key(wallet_address_factory):
     """Test the initialization of a WalletAddress with a key."""
+    wallet_address_with_key = wallet_address_factory(key=True)
+
     assert wallet_address_with_key.network_id == "base-sepolia"
     assert wallet_address_with_key.address_id == "0x1234567890123456789012345678901234567890"
-    assert wallet_address_with_key.wallet_id == "test-wallet-id-1"
+    assert wallet_address_with_key.wallet_id == "test-wallet-id"
     assert wallet_address_with_key.key is not None
     assert wallet_address_with_key.can_sign
 
 
-def test_wallet_id_property(wallet_address):
+def test_wallet_id_property(wallet_address_factory):
     """Test the wallet_id property of a WalletAddress."""
-    assert wallet_address.wallet_id == "test-wallet-id-1"
+    wallet_address = wallet_address_factory()
+
+    assert wallet_address.wallet_id == "test-wallet-id"
 
 
-def test_key_property(wallet_address, wallet_address_with_key):
+def test_key_property(wallet_address_factory):
     """Test the key property of a WalletAddress."""
+    wallet_address = wallet_address_factory()
+    wallet_address_with_key = wallet_address_factory(True)
+
     assert wallet_address.key is None
     assert wallet_address_with_key.key is not None
 
 
-def test_key_setter(wallet_address):
+def test_key_setter(wallet_address_factory):
     """Test the key setter of a WalletAddress."""
+    wallet_address = wallet_address_factory()
+
     new_key = Mock(spec=LocalAccount)
     wallet_address.key = new_key
     assert wallet_address.key == new_key
 
 
-def test_can_sign_property(wallet_address, wallet_address_with_key):
+def test_can_sign_property(wallet_address_factory):
     """Test the can_sign property of a WalletAddress."""
+    wallet_address = wallet_address_factory()
+    wallet_address_with_key = wallet_address_factory(True)
+
     assert not wallet_address.can_sign
     assert wallet_address_with_key.can_sign
 
 
-def test_key_setter_raises_error_when_already_set(wallet_address_with_key):
+def test_key_setter_raises_error_when_already_set(wallet_address_factory):
     """Test the key setter raises an error when already set."""
+    wallet_address_with_key = wallet_address_factory(True)
+
     new_key = Mock(spec=LocalAccount)
     with pytest.raises(ValueError, match="Private key is already set"):
         wallet_address_with_key.key = new_key
@@ -115,9 +90,14 @@ def test_key_setter_raises_error_when_already_set(wallet_address_with_key):
 @patch("cdp.Cdp.api_clients")
 @patch("cdp.Cdp.use_server_signer", True)
 def test_transfer_with_server_signer(
-    mock_api_clients, mock_transfer, wallet_address, balance_model
+    mock_api_clients, mock_transfer, wallet_address_factory, balance_model_factory
 ):
     """Test the transfer method with a server signer."""
+    wallet_address = wallet_address_factory()
+    balance_model = balance_model_factory(
+        amount="5000000000000000000", network_id="base-sepolia", asset_id="eth", decimals=18
+    )
+
     mock_transfer_instance = Mock(spec=Transfer)
     mock_transfer.create.return_value = mock_transfer_instance
 
@@ -147,8 +127,13 @@ def test_transfer_with_server_signer(
 @patch("cdp.wallet_address.Transfer")
 @patch("cdp.Cdp.api_clients")
 @patch("cdp.Cdp.use_server_signer", False)
-def test_transfer(mock_api_clients, mock_transfer, wallet_address_with_key, balance_model):
+def test_transfer(mock_api_clients, mock_transfer, wallet_address_factory, balance_model_factory):
     """Test the transfer method."""
+    wallet_address_with_key = wallet_address_factory(key=True)
+    balance_model = balance_model_factory(
+        amount="5000000000000000000", network_id="base-sepolia", asset_id="eth", decimals=18
+    )
+
     mock_transfer_instance = Mock(spec=Transfer)
     mock_transfer.create.return_value = mock_transfer_instance
 
@@ -183,9 +168,14 @@ def test_transfer(mock_api_clients, mock_transfer, wallet_address_with_key, bala
 @patch("cdp.Cdp.api_clients")
 @patch("cdp.Cdp.use_server_signer", False)
 def test_transfer_create_api_error(
-    mock_api_clients, mock_transfer, wallet_address_with_key, balance_model
+    mock_api_clients, mock_transfer, wallet_address_factory, balance_model_factory
 ):
     """Test the transfer method raises an error when the create API call fails."""
+    wallet_address_with_key = wallet_address_factory(key=True)
+    balance_model = balance_model_factory(
+        amount="5000000000000000000", network_id="base-sepolia", asset_id="eth", decimals=18
+    )
+
     mock_transfer.create.side_effect = Exception("API Error")
 
     mock_get_balance = Mock()
@@ -215,9 +205,14 @@ def test_transfer_create_api_error(
 @patch("cdp.Cdp.api_clients")
 @patch("cdp.Cdp.use_server_signer", False)
 def test_transfer_broadcast_api_error(
-    mock_api_clients, mock_transfer, wallet_address_with_key, balance_model
+    mock_api_clients, mock_transfer, wallet_address_factory, balance_model_factory
 ):
     """Test the transfer method raises an error when the broadcast API call fails."""
+    wallet_address_with_key = wallet_address_factory(key=True)
+    balance_model = balance_model_factory(
+        amount="5000000000000000000", network_id="base-sepolia", asset_id="eth", decimals=18
+    )
+
     mock_transfer_instance = Mock(spec=Transfer)
     mock_transfer_instance.broadcast.side_effect = Exception("API Error")
     mock_transfer.create.return_value = mock_transfer_instance
@@ -250,8 +245,13 @@ def test_transfer_broadcast_api_error(
 @patch("cdp.wallet_address.Trade")
 @patch("cdp.Cdp.api_clients")
 @patch("cdp.Cdp.use_server_signer", True)
-def test_trade(mock_api_clients, mock_trade, wallet_address, balance_model):
+def test_trade(mock_api_clients, mock_trade, wallet_address_factory, balance_model_factory):
     """Test the trade method with a server signer."""
+    wallet_address = wallet_address_factory()
+    balance_model = balance_model_factory(
+        amount="5000000000000000000", network_id="base-sepolia", asset_id="eth", decimals=18
+    )
+
     mock_trade.create.return_value = Mock(spec=Trade)
 
     mock_get_balance = Mock()
@@ -275,9 +275,14 @@ def test_trade(mock_api_clients, mock_trade, wallet_address, balance_model):
 @patch("cdp.Cdp.api_clients")
 @patch("cdp.Cdp.use_server_signer", False)
 def test_trade_with_client_signer(
-    mock_api_clients, mock_trade, wallet_address_with_key, balance_model
+    mock_api_clients, mock_trade, wallet_address_factory, balance_model_factory
 ):
     """Test the trade method."""
+    wallet_address_with_key = wallet_address_factory(key=True)
+    balance_model = balance_model_factory(
+        amount="5000000000000000000", network_id="base-sepolia", asset_id="eth", decimals=18
+    )
+
     mock_trade_instance = Mock(spec=Trade)
     mock_trade.create.return_value = mock_trade_instance
 
@@ -301,8 +306,10 @@ def test_trade_with_client_signer(
 
 
 @patch("cdp.wallet_address.Transfer")
-def test_transfers(mock_transfer, wallet_address):
+def test_transfers(mock_transfer, wallet_address_factory):
     """Test the transfers method."""
+    wallet_address = wallet_address_factory()
+
     mock_transfer.list.return_value = iter([Mock(spec=Transfer), Mock(spec=Transfer)])
 
     transfers = wallet_address.transfers()
@@ -315,8 +322,10 @@ def test_transfers(mock_transfer, wallet_address):
 
 
 @patch("cdp.wallet_address.Transfer")
-def test_transfers_api_error(mock_transfer, wallet_address):
+def test_transfers_api_error(mock_transfer, wallet_address_factory):
     """Test the transfers method raises an error when the API call fails."""
+    wallet_address = wallet_address_factory()
+
     mock_transfer.list.side_effect = Exception("API Error")
 
     with pytest.raises(Exception, match="API Error"):
@@ -324,8 +333,10 @@ def test_transfers_api_error(mock_transfer, wallet_address):
 
 
 @patch("cdp.wallet_address.Trade")
-def test_trades(mock_trade, wallet_address):
+def test_trades(mock_trade, wallet_address_factory):
     """Test the trades method."""
+    wallet_address = wallet_address_factory()
+
     mock_trade.list.return_value = iter([Mock(spec=Trade), Mock(spec=Trade)])
 
     trades = wallet_address.trades()
@@ -338,8 +349,10 @@ def test_trades(mock_trade, wallet_address):
 
 
 @patch("cdp.wallet_address.Trade")
-def test_trades_api_error(mock_trade, wallet_address):
+def test_trades_api_error(mock_trade, wallet_address_factory):
     """Test the trades method raises an error when the API call fails."""
+    wallet_address = wallet_address_factory()
+
     mock_trade.list.side_effect = Exception("API Error")
 
     with pytest.raises(Exception, match="API Error"):
@@ -350,9 +363,14 @@ def test_trades_api_error(mock_trade, wallet_address):
 @patch("cdp.Cdp.api_clients")
 @patch("cdp.Cdp.use_server_signer", True)
 def test_invoke_contract_with_server_signer(
-    mock_api_clients, mock_contract_invocation, wallet_address, balance_model
+    mock_api_clients, mock_contract_invocation, wallet_address_factory, balance_model_factory
 ):
     """Test the invoke_contract method with a server signer."""
+    wallet_address = wallet_address_factory()
+    balance_model = balance_model_factory(
+        amount="5000000000000000000", network_id="base-sepolia", asset_id="eth", decimals=18
+    )
+
     mock_contract_invocation_instance = Mock(spec=ContractInvocation)
     mock_contract_invocation.create.return_value = mock_contract_invocation_instance
 
@@ -392,9 +410,14 @@ def test_invoke_contract_with_server_signer(
 @patch("cdp.Cdp.api_clients")
 @patch("cdp.Cdp.use_server_signer", False)
 def test_invoke_contract(
-    mock_api_clients, mock_contract_invocation, wallet_address_with_key, balance_model
+    mock_api_clients, mock_contract_invocation, wallet_address_factory, balance_model_factory
 ):
     """Test the invoke_contract method."""
+    wallet_address_with_key = wallet_address_factory(key=True)
+    balance_model = balance_model_factory(
+        amount="5000000000000000000", network_id="base-sepolia", asset_id="eth", decimals=18
+    )
+
     mock_contract_invocation_instance = Mock(spec=ContractInvocation)
     mock_contract_invocation.create.return_value = mock_contract_invocation_instance
 
@@ -436,9 +459,14 @@ def test_invoke_contract(
 @patch("cdp.Cdp.api_clients")
 @patch("cdp.Cdp.use_server_signer", False)
 def test_invoke_contract_api_error(
-    mock_api_clients, mock_contract_invocation, wallet_address_with_key, balance_model
+    mock_api_clients, mock_contract_invocation, wallet_address_factory, balance_model_factory
 ):
     """Test the invoke_contract method raises an error when the create API call fails."""
+    wallet_address_with_key = wallet_address_factory(key=True)
+    balance_model = balance_model_factory(
+        amount="5000000000000000000", network_id="base-sepolia", asset_id="eth", decimals=18
+    )
+
     mock_contract_invocation.create.side_effect = Exception("API Error")
 
     mock_get_balance = Mock()
@@ -477,9 +505,14 @@ def test_invoke_contract_api_error(
 @patch("cdp.Cdp.api_clients")
 @patch("cdp.Cdp.use_server_signer", False)
 def test_invoke_contract_broadcast_api_error(
-    mock_api_clients, mock_contract_invocation, wallet_address_with_key, balance_model
+    mock_api_clients, mock_contract_invocation, wallet_address_factory, balance_model_factory
 ):
     """Test the invoke_contract method raises an error when the broadcast API call fails."""
+    wallet_address_with_key = wallet_address_factory(key=True)
+    balance_model = balance_model_factory(
+        amount="5000000000000000000", network_id="base-sepolia", asset_id="eth", decimals=18
+    )
+
     mock_contract_invocation_instance = Mock(spec=ContractInvocation)
     mock_contract_invocation.create.return_value = mock_contract_invocation_instance
     mock_contract_invocation_instance.broadcast.side_effect = Exception("API Error")
@@ -520,8 +553,10 @@ def test_invoke_contract_broadcast_api_error(
 
 @patch("cdp.wallet_address.PayloadSignature")
 @patch("cdp.Cdp.use_server_signer", True)
-def test_sign_payload_with_server_signer(mock_payload_signature, wallet_address):
+def test_sign_payload_with_server_signer(mock_payload_signature, wallet_address_factory):
     """Test the sign_payload method with a server signer."""
+    wallet_address = wallet_address_factory()
+
     mock_payload_signature_instance = Mock(spec=PayloadSignature)
     mock_payload_signature.create.return_value = mock_payload_signature_instance
 
@@ -539,8 +574,10 @@ def test_sign_payload_with_server_signer(mock_payload_signature, wallet_address)
 @patch("cdp.wallet_address.to_hex", Mock(return_value="0xsignature"))
 @patch("cdp.wallet_address.PayloadSignature")
 @patch("cdp.Cdp.use_server_signer", False)
-def test_sign_payload(mock_payload_signature, wallet_address_with_key):
+def test_sign_payload(mock_payload_signature, wallet_address_factory):
     """Test the sign_payload method."""
+    wallet_address_with_key = wallet_address_factory(key=True)
+
     mock_payload_signature_instance = Mock(spec=PayloadSignature)
     mock_payload_signature.create.return_value = mock_payload_signature_instance
 
@@ -568,8 +605,10 @@ def test_sign_payload(mock_payload_signature, wallet_address_with_key):
 @patch("cdp.wallet_address.to_hex", Mock(return_value="0xsignature"))
 @patch("cdp.wallet_address.PayloadSignature")
 @patch("cdp.Cdp.use_server_signer", False)
-def test_sign_payload_api_error(mock_payload_signature, wallet_address_with_key):
+def test_sign_payload_api_error(mock_payload_signature, wallet_address_factory):
     """Test the sign_payload method."""
+    wallet_address_with_key = wallet_address_factory(key=True)
+
     mock_signature = Mock(spec=SignedMessage)
     mock_signature.signature = "0xsignature"
     wallet_address_with_key.key.unsafe_sign_hash.return_value = mock_signature
@@ -594,8 +633,15 @@ def test_sign_payload_api_error(mock_payload_signature, wallet_address_with_key)
 
 
 @patch("cdp.Cdp.api_clients")
-def test_ensure_sufficient_balance_sufficient(mock_api_clients, wallet_address, balance_model):
+def test_ensure_sufficient_balance_sufficient(
+    mock_api_clients, wallet_address_factory, balance_model_factory
+):
     """Test the ensure_sufficient_balance method with sufficient balance."""
+    wallet_address = wallet_address_factory()
+    balance_model = balance_model_factory(
+        amount="5000000000000000000", network_id="base-sepolia", asset_id="eth", decimals=18
+    )
+
     mock_get_balance = Mock()
     mock_get_balance.return_value = balance_model
     mock_api_clients.external_addresses.get_external_address_balance = mock_get_balance
@@ -608,8 +654,15 @@ def test_ensure_sufficient_balance_sufficient(mock_api_clients, wallet_address, 
 
 
 @patch("cdp.Cdp.api_clients")
-def test_ensure_sufficient_balance_insufficient(mock_api_clients, wallet_address, balance_model):
+def test_ensure_sufficient_balance_insufficient(
+    mock_api_clients, wallet_address_factory, balance_model_factory
+):
     """Test the ensure_sufficient_balance method with insufficient balance."""
+    wallet_address = wallet_address_factory()
+    balance_model = balance_model_factory(
+        amount="5000000000000000000", network_id="base-sepolia", asset_id="eth", decimals=18
+    )
+
     mock_get_balance = Mock()
     mock_get_balance.return_value = balance_model
     mock_api_clients.external_addresses.get_external_address_balance = mock_get_balance
@@ -622,22 +675,28 @@ def test_ensure_sufficient_balance_insufficient(mock_api_clients, wallet_address
     )
 
 
-def test_str_representation(wallet_address):
+def test_str_representation(wallet_address_factory):
     """Test the str representation of a WalletAddress."""
-    expected_str = "WalletAddress: (address_id: 0x1234567890123456789012345678901234567890, wallet_id: test-wallet-id-1, network_id: base-sepolia)"
+    wallet_address = wallet_address_factory()
+
+    expected_str = "WalletAddress: (address_id: 0x1234567890123456789012345678901234567890, wallet_id: test-wallet-id, network_id: base-sepolia)"
     assert str(wallet_address) == expected_str
 
 
-def test_repr(wallet_address):
+def test_repr(wallet_address_factory):
     """Test the repr representation of a WalletAddress."""
-    expected_repr = "WalletAddress: (address_id: 0x1234567890123456789012345678901234567890, wallet_id: test-wallet-id-1, network_id: base-sepolia)"
+    wallet_address = wallet_address_factory()
+
+    expected_repr = "WalletAddress: (address_id: 0x1234567890123456789012345678901234567890, wallet_id: test-wallet-id, network_id: base-sepolia)"
     assert repr(wallet_address) == expected_repr
 
 
 @patch("cdp.Cdp.api_clients")
 @patch("cdp.Cdp.use_server_signer", True)
-def test_wallet_address_deploy_token_total_supply_string(mock_api_clients, wallet_address):
+def test_wallet_address_deploy_token_total_supply_string(mock_api_clients, wallet_address_factory):
     """Test the deploy_token method of a WalletAddress with a string total_supply."""
+    wallet_address = wallet_address_factory()
+
     mock_smart_contract = Mock(spec=SmartContract)
     mock_api_clients.smart_contracts.create_smart_contract.return_value = mock_smart_contract
 
@@ -662,8 +721,10 @@ def test_wallet_address_deploy_token_total_supply_string(mock_api_clients, walle
 
 @patch("cdp.Cdp.api_clients")
 @patch("cdp.Cdp.use_server_signer", True)
-def test_wallet_address_deploy_token_total_supply_number(mock_api_clients, wallet_address):
+def test_wallet_address_deploy_token_total_supply_number(mock_api_clients, wallet_address_factory):
     """Test the deploy_token method of a WalletAddress with a number total_supply."""
+    wallet_address = wallet_address_factory()
+
     mock_smart_contract = Mock(spec=SmartContract)
     mock_api_clients.smart_contracts.create_smart_contract.return_value = mock_smart_contract
 
@@ -688,8 +749,10 @@ def test_wallet_address_deploy_token_total_supply_number(mock_api_clients, walle
 
 @patch("cdp.Cdp.api_clients")
 @patch("cdp.Cdp.use_server_signer", True)
-def test_wallet_address_deploy_token_total_supply_decimal(mock_api_clients, wallet_address):
+def test_wallet_address_deploy_token_total_supply_decimal(mock_api_clients, wallet_address_factory):
     """Test the deploy_token method of a WalletAddress with a Decimal total_supply."""
+    wallet_address = wallet_address_factory()
+
     mock_smart_contract = Mock(spec=SmartContract)
     mock_api_clients.smart_contracts.create_smart_contract.return_value = mock_smart_contract
 
@@ -714,8 +777,10 @@ def test_wallet_address_deploy_token_total_supply_decimal(mock_api_clients, wall
 
 @patch("cdp.Cdp.api_clients")
 @patch("cdp.Cdp.use_server_signer", True)
-def test_wallet_address_deploy_nft(mock_api_clients, wallet_address):
+def test_wallet_address_deploy_nft(mock_api_clients, wallet_address_factory):
     """Test the deploy_nft method of a WalletAddress."""
+    wallet_address = wallet_address_factory()
+
     mock_smart_contract = Mock(spec=SmartContract)
     mock_api_clients.smart_contracts.create_smart_contract.return_value = mock_smart_contract
 
@@ -740,8 +805,10 @@ def test_wallet_address_deploy_nft(mock_api_clients, wallet_address):
 
 @patch("cdp.Cdp.api_clients")
 @patch("cdp.Cdp.use_server_signer", True)
-def test_wallet_address_deploy_multi_token(mock_api_clients, wallet_address):
+def test_wallet_address_deploy_multi_token(mock_api_clients, wallet_address_factory):
     """Test the deploy_multi_token method of a WalletAddress."""
+    wallet_address = wallet_address_factory()
+
     mock_smart_contract = Mock(spec=SmartContract)
     mock_api_clients.smart_contracts.create_smart_contract.return_value = mock_smart_contract
 
@@ -766,8 +833,10 @@ def test_wallet_address_deploy_multi_token(mock_api_clients, wallet_address):
 
 @patch("cdp.Cdp.api_clients")
 @patch("cdp.Cdp.use_server_signer", True)
-def test_wallet_address_deploy_token_with_server_signer(mock_api_clients, wallet_address):
+def test_wallet_address_deploy_token_with_server_signer(mock_api_clients, wallet_address_factory):
     """Test the deploy_token method of a WalletAddress with server signer."""
+    wallet_address = wallet_address_factory()
+
     mock_smart_contract = Mock(spec=SmartContract)
     mock_api_clients.smart_contracts.create_smart_contract.return_value = mock_smart_contract
 
@@ -795,8 +864,10 @@ def test_wallet_address_deploy_token_with_server_signer(mock_api_clients, wallet
 
 @patch("cdp.Cdp.api_clients")
 @patch("cdp.Cdp.use_server_signer", True)
-def test_wallet_address_deploy_nft_with_server_signer(mock_api_clients, wallet_address):
+def test_wallet_address_deploy_nft_with_server_signer(mock_api_clients, wallet_address_factory):
     """Test the deploy_nft method of a WalletAddress with server signer."""
+    wallet_address = wallet_address_factory()
+
     mock_smart_contract = Mock(spec=SmartContract)
     mock_api_clients.smart_contracts.create_smart_contract.return_value = mock_smart_contract
 
@@ -824,8 +895,12 @@ def test_wallet_address_deploy_nft_with_server_signer(mock_api_clients, wallet_a
 
 @patch("cdp.Cdp.api_clients")
 @patch("cdp.Cdp.use_server_signer", True)
-def test_wallet_address_deploy_multi_token_with_server_signer(mock_api_clients, wallet_address):
+def test_wallet_address_deploy_multi_token_with_server_signer(
+    mock_api_clients, wallet_address_factory
+):
     """Test the deploy_multi_token method of a WalletAddress with server signer."""
+    wallet_address = wallet_address_factory()
+
     mock_smart_contract = Mock(spec=SmartContract)
     mock_api_clients.smart_contracts.create_smart_contract.return_value = mock_smart_contract
 
@@ -852,8 +927,10 @@ def test_wallet_address_deploy_multi_token_with_server_signer(mock_api_clients, 
 
 
 @patch("cdp.wallet_address.SmartContract")
-def test_deploy_token_api_error(mock_smart_contract, wallet_address_with_key):
+def test_deploy_token_api_error(mock_smart_contract, wallet_address_factory):
     """Test the deploy_token method raises an error when the create API call fails."""
+    wallet_address_with_key = wallet_address_factory(key=True)
+
     mock_smart_contract.create.side_effect = Exception("API Error")
 
     with pytest.raises(Exception, match="API Error"):
@@ -863,8 +940,10 @@ def test_deploy_token_api_error(mock_smart_contract, wallet_address_with_key):
 
 
 @patch("cdp.wallet_address.SmartContract")
-def test_deploy_token_broadcast_api_error(mock_smart_contract, wallet_address_with_key):
+def test_deploy_token_broadcast_api_error(mock_smart_contract, wallet_address_factory):
     """Test the deploy_token method raises an error when the broadcast API call fails."""
+    wallet_address_with_key = wallet_address_factory(key=True)
+
     mock_smart_contract_instance = Mock(spec=SmartContract)
     mock_smart_contract.create.return_value = mock_smart_contract_instance
     mock_smart_contract_instance.broadcast.side_effect = Exception("API Error")
@@ -878,8 +957,10 @@ def test_deploy_token_broadcast_api_error(mock_smart_contract, wallet_address_wi
 
 
 @patch("cdp.wallet_address.SmartContract")
-def test_deploy_nft_api_error(mock_smart_contract, wallet_address_with_key):
+def test_deploy_nft_api_error(mock_smart_contract, wallet_address_factory):
     """Test the deploy_nft method raises an error when the create API call fails."""
+    wallet_address_with_key = wallet_address_factory(key=True)
+
     mock_smart_contract.create.side_effect = Exception("API Error")
 
     with pytest.raises(Exception, match="API Error"):
@@ -891,8 +972,10 @@ def test_deploy_nft_api_error(mock_smart_contract, wallet_address_with_key):
 
 
 @patch("cdp.wallet_address.SmartContract")
-def test_deploy_nft_broadcast_api_error(mock_smart_contract, wallet_address_with_key):
+def test_deploy_nft_broadcast_api_error(mock_smart_contract, wallet_address_factory):
     """Test the deploy_nft method raises an error when the broadcast API call fails."""
+    wallet_address_with_key = wallet_address_factory(key=True)
+
     mock_smart_contract_instance = Mock(spec=SmartContract)
     mock_smart_contract.create.return_value = mock_smart_contract_instance
     mock_smart_contract_instance.broadcast.side_effect = Exception("API Error")
@@ -908,8 +991,10 @@ def test_deploy_nft_broadcast_api_error(mock_smart_contract, wallet_address_with
 
 
 @patch("cdp.wallet_address.SmartContract")
-def test_deploy_multi_token_api_error(mock_smart_contract, wallet_address_with_key):
+def test_deploy_multi_token_api_error(mock_smart_contract, wallet_address_factory):
     """Test the deploy_multi_token method raises an error when the create API call fails."""
+    wallet_address_with_key = wallet_address_factory(key=True)
+
     mock_smart_contract.create.side_effect = Exception("API Error")
 
     with pytest.raises(Exception, match="API Error"):
@@ -919,8 +1004,10 @@ def test_deploy_multi_token_api_error(mock_smart_contract, wallet_address_with_k
 
 
 @patch("cdp.wallet_address.SmartContract")
-def test_deploy_multi_token_broadcast_api_error(mock_smart_contract, wallet_address_with_key):
+def test_deploy_multi_token_broadcast_api_error(mock_smart_contract, wallet_address_factory):
     """Test the deploy_multi_token method raises an error when the broadcast API call fails."""
+    wallet_address_with_key = wallet_address_factory(key=True)
+
     mock_smart_contract_instance = Mock(spec=SmartContract)
     mock_smart_contract.create.return_value = mock_smart_contract_instance
     mock_smart_contract_instance.broadcast.side_effect = Exception("API Error")
